Refactor tools monitor runtime into modular files
